Rename TypeUtilities to TypeExtensions and move it to common\extensions
